requirements.txt
flask==2.0.1
pandas==1.5.3
plotly==5.18.0
langchain==0.0.350
openai>=1.6.1
python-dotenv==1.0.0
chromadb==0.4.18
